delete

+++ app/helpers.py
@@ -191,15 +191,21 @@ def _get_files_to_not_delete():
     return not_delete
 
 
-def _remove_file_older_than(file_path: str, max_age_limit: float):
-    # If the file is older than the age limit, delete it
-    if os.path.getmtime(file_path) < max_age_limit:
-        try:
-            log.info(f"Deleting {file_path}")
-            os.remove(file_path)
-        except OSError as e:
-            log.info(f"Error: Could not delete {file_path}. Reason: {e.strerror}")
+# def _remove_file_older_than(file_path: str, max_age_limit: float):
+#     # If the file is older than the age limit, delete it
+#     if os.path.getmtime(file_path) < max_age_limit:
+#         try:
+#             log.info(f"Deleting {file_path}")
+#             os.remove(file_path)
+#         except OSError as e:
+#             log.info(f"Error: Could not delete {file_path}. Reason: {e.strerror}")
 
+def _remove_file_older_than(file_path: str, max_age_limit: float):
+    if os.path.exists(file_path):  # 检查文件是否存在
+        if os.path.getmtime(file_path) < max_age_limit:
+            os.remove(file_path)
+    else:
+        print(f"文件 {file_path} 不存在，跳过删除。")
 
 def delete_old_files(directory: str, age_limit_seconds: int):
     files_to_not_delete = _get_files_to_not_delete()
